* for 64bit.
git-svn-id: http://svn.sourceforge.jp/svnroot/modchxj/mod_chxj/branches/RELEASE_0_11_0@3188 1a406e8e-add9-4483-a2c8-d8cac5b7c224

+++ src/chxj_mysql.c
@@ -1121,7 +1121,7 @@ int
 chxj_cookie_unlock_mysql(request_rec *r, mod_chxj_config *m)
 {
   char query[MAX_STRING_LEN];
-  DBG(r, "REQ[%X] start chxj_cookie_unlock_mysql()", (apr_size_t)r);
+  DBG(r, "REQ[%X] start chxj_cookie_unlock_mysql()", (unsigned int)(apr_size_t)r);
   if (! chxj_open_mysql_handle(r, m)) {
     ERR(r, "Cannot open mysql connection");
     DBG(r, "end   chxj_save_cookie_expire_mysql()");
